requirements.txt
alembic==1.11.1
APScheduler==3.10.1
backports.zoneinfo==0.2.1
bcrypt==4.0.1
blinker==1.6.2
cffi==1.15.1
click==8.1.3
cryptography==41.0.1
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
mysqlclient==2.1.1
paramiko==3.2.0
protobuf==3.20.3
pycparser==2.21
PyJWT==2.7.0
PyNaCl==1.5.0
pysftp==0.2.9
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.16
typing-extensions==4.6.3
tzlocal==5.0.1
Werkzeug==2.3.6
zipp==3.15.0
